requirements.txt
requests
python-dotenv
slack-sdk
pandas
plotly
reportlab
watchdog
scapy
yara-python
scikit-learn
radare2
pysandbox

#py -m pip install requests python-dotenv slack-sdk pandas plotly reportlab watchdog scapy yara-python scikit-learn radare2 pysandbox